pyproject.toml
[project]
name = "murmur"
version = "0.1.0"
description = "A social media showcase project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cotton>=2.1.3",
    "django-ninja>=1.4.3",
    "django-ninja-extra>=0.30.1",
    "django-ninja-jwt[crypto]>=5.3.7",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.1",
    "django-stubs-ext>=5.2.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.murmur.settings"
python_files = ["test_*.py", "*_test.py", "tests/*.py", "tests.py"]

[tool.pyright]
venv = ".venv"
venvPath = "."
